+++ src/test/java/com/ttasjwi/cafe/controller/UserControllerTest.java
@@ -9,8 +9,8 @@ import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.test.web.servlet.MockMvc;
 
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.view;
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;
 
 @WebMvcTest(controllers = UserController.class)
 class UserControllerTest {
@@ -22,13 +22,26 @@ class UserControllerTest {
     private UserRepository userRepository;
 
     @Test
-    @DisplayName("\"/users/new\"로 요청하면 \"/users/createUserForm\" 이 반환된다.")
+    @DisplayName("\"/users/new\"로 GET 요청하면 \"/users/createUserForm\" 이 반환된다.")
     void createFormTest() throws Exception {
         String requestUrl = "/users/new";
 
         mockMvc.perform(get(requestUrl))
                 .andExpect(status().isOk())
                 .andExpect(view().name("/users/createUserForm"));
+    }
+
+    @Test
+    @DisplayName("\"/users/new\"로 POST 요청하면 \"/users\" 로 리다이렉트 된다.")
+    void createTest() throws Exception {
+        String requestUrl = "/users/new";
+
+        mockMvc.perform(post(requestUrl)
+                        .param("userName","ttasjwi")
+                        .param("userEmail","ttasjwi920@gmail.com")
+                        .param("password","1234"))
+                .andExpect(status().is3xxRedirection())
+                .andExpect(redirectedUrl("/users"));
     }
 
     @Test